backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.12.0
numpy==1.26.3
websockets==12.0
pydantic==2.6.1
whisper==1.1.10
python-multipart==0.0.6 